pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flowchem"
version = "1.1.0-post1"

description = "Flowchem is a library to control instruments and devices commonly found in chemistry labs via an interoperable web API."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
license-files = ['LICEN[CS]E*',]
keywords = ["chemistry", "automation", "laboratory", "science"]
authors = [
    { name = "Dario Cambié", email = "2422614+dcambie@users.noreply.github.com" },
    { name = "Jakob Wolf", email = "75418671+JB-Wolf@users.noreply.github.com" },
    { name = "Wei-Hsin Hsu", email = "98024191+cynthiabour@users.noreply.github.com" },
    { name = "Samuel Saraiva", email = "88909409+samuelvitorsaraiva@users.noreply.github.com" },
]
maintainers = [
    { name = "Jakob Wolf", email = "75418671+JB-Wolf@users.noreply.github.com" },
    { name = "Samuel Saraiva", email = "88909409+samuelvitorsaraiva@users.noreply.github.com" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Development Status :: 5 - Production/Stable",
    "Operating System :: OS Independent"
]
dependencies = [
    "aioserial>=1.3.1",
    "anyio",
    "asyncua>=1.0.2",
    "bronkhorst-propar>=1.1.0",
    "click<=8.1.3", # Temporary due to https://github.com/pallets/click/issues/2558
    "fastapi>=0.100.0",
    "ifaddr>=0.2.0",
    "loguru>=0.7.0",
    "lxml>=4.9.2",
    "numpy",
    "packaging>=23.1",
    "pint>=0.16.1,!=0.21", # See hgrecco/pint#1642
    "pydantic[email]>=2.0.2",
    "pyserial>=3",
    "requests",
    "rich_click>=1.6.1",
    "tomli; python_version < \"3.11\"",
    "uvicorn>=0.19.0",
    "zeroconf>=0.71.0",
    "python-multipart>=0.0.20",
]

[project.optional-dependencies]
all = [
    "black",
    "data-science-types",
    "lxml-stubs",
    "mypy",
    "pre-commit",
    "ruff>=0.0.252",
    "types-lxml",
    "types-PyYAML",
    "types-requests",
    "flowchem-test>=0.1b4",
    "httpx",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-mock",
    "pytest-xprocess",
    "phidget22>=1.7.20211005",
    "furo",
    "mistune==0.8.4",
    "myst-parser",
    "sphinx",
    "sphinx-design",
    "sphinx-autodoc-typehints",
    "sphinx-rtd-theme",
    "sphinxcontrib-openapi"
]
ci = [
    "black",
    "data-science-types",
    "lxml-stubs",
    "mypy",
    "pre-commit",
    "ruff>=0.0.252",
    "types-lxml",
    "types-PyYAML",
    "types-requests",
    "flowchem-test>=0.1b4",
    "httpx",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-mock",
    "pytest-xprocess",
    "furo",
    "mistune==0.8.4",
    "myst-parser",
    "sphinx",
    "sphinx-design",
    "sphinx-autodoc-typehints",
    "sphinx-rtd-theme",
    "sphinxcontrib-openapi"
]
dev = [
    "black",
    "data-science-types",
    "lxml-stubs",
    "mypy",
    "pre-commit",
    "ruff>=0.0.252",
    "types-lxml",
    "types-PyYAML",
    "types-requests",
]
test = [
    "flowchem-test>=0.1b4",
    "httpx",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "pytest-mock",
    "pytest-xprocess",
]
phidget = [
    "phidget22>=1.7.20211005",
]
docs = [
    "furo",
    "mistune==0.8.4", # Due to sphinx-contrib/openapi#121
    "myst-parser",
    "sphinx",
    "sphinx-design",
    "sphinx-autodoc-typehints",
    "sphinx-rtd-theme",
    "sphinxcontrib-openapi",
]

[project.urls]
homepage = "https://github.com/cambiegroup/flowchem"
documentation = "https://flowchem.readthedocs.io"
repository = "https://github.com/cambiegroup/flowchem"

[project.scripts]
flowchem = "flowchem.__main__:main"
flowchem-autodiscover = "flowchem.utils.device_finder:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
flowchem = ["py.typed"]

[tool.mypy]
ignore_missing_imports = true
python_version = "3.11"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
# Note: pytest cov is not compatible with the pycharm debugger in tests
addopts = "-m 'not HApump and not Spinsolve and not FlowIR and not KPump and not KPump and not ml600 and not Knauer_Distribution_Valve and not CVC3000 and not Manson' --no-cov"
markers = [
    "HApump: tests requiring a local HA Elite11 connected.",
    "Spinsolve: tests requiring a connection to Spinsolve.",
    "FlowIR: tests requiring a connection to a FlowIR.",
    "KPump: tests for Azura compact.",
    "ml600: test for ML600 pump",
    "Knauer_Distribution_Valve: test knauer distribution valve",
    "CVC3000: Test CVC3000 vacun pump",
    "Manson: Test Manson power control"
]

[tool.ruff]
line-length = 120
# Allow imports relative to the "src" and "test" directories.
src = ["src", "test"]
lint.per-file-ignores."__init__.py" = ["F403"]

